packages/tangram_weather/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tangram_weather"
version = "0.2.0"
readme = "readme.md"
dependencies = [
    "tangram_core>=0.2.0",
    "xarray>=2025.6.1",
    "cfgrib>=0.9.15.0",
    "tqdm>=4.67.1",
    "orjson>=3.10.18",
    "pandas",
    "numpy>=2",
    "pillow>=10",
]
authors = [
    { name = "Xavier Olive", email = "git@xoolive.org" }
]

[project.entry-points."tangram_core.plugins"]
tangram_weather = "tangram_weather:plugin"

[tool.uv.sources]
tangram_core = { workspace = true }

[tool.hatch.build.targets.sdist]
ignore-vcs = true
include = ["dist-frontend/*", "src/*"]

[tool.hatch.build.targets.wheel.force-include]
"dist-frontend" = "tangram_weather/dist-frontend"
